+++ CHANGELOG.md
@@ -159,6 +159,7 @@ with the `YY.MINOR.MICRO` scheme.
 - Deprecated `plugin_dir` setting is removed following a major release. Use `share_dir` instead.
 - Deprecated `advertise_address` setting is removed following a major release. Use `iproto` section instead.
 - `--cluster-name` argument for `picodata expel` is removed following a major release as it is no longer needed.
+- Deprecated `ServiceWorkerManager` is removed from `picodata_plugin` following a major release.
 
 ## [26.1.1] - 2026-03-24